LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-length-value-dynamic-expected.txt


FAIL Initially, the sibling-index() is 2 for #target assert_equals: expected "200px" but got "13px"
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "100px" but got "13px"
